# Conflicts:
#	app/src/main/java/com/example/mishlavim/model/Firebase/FirestoreMethods.java

+++ app/src/main/java/com/example/mishlavim/guideActivities/GuideMainActivity.java
@@ -121,25 +121,30 @@ public class GuideMainActivity extends AppCompatActivity implements View.OnClick
 
     @Override
     public boolean onMenuItemClick(MenuItem item) {
-        switch (item.getItemId()){
         //delete volunteer
-            case R.id.remove_volunteer:
+            if (item.getItemId() == R.id.remove_volunteer) {
                 DialogFragment newFragment = new DeleteUser();
                 newFragment.show(getSupportFragmentManager(), "deleteUser");
                 return true;
-
-            case R.id.view_volunteer:
-                Log.d("clicked:", clickedRowName + " view" );
-                return true;
-            case R.id.edit_volunteer:
-                Intent intent = new Intent(getApplicationContext(), GuideVoluSettingActivity.class);
-                intent.putExtra("CLICKED_VOLU_KEY", clickedRowName);
-                intent.putExtra("CLICKED_VOLU_ID", guide.getMyVolunteers().get(clickedRowName));
-                startActivity(intent);
-                overridePendingTransition(0, 0);
-                return true;
             }
-            return false;
+
+        else if (item.getItemId() == R.id.edit_volunteer) {
+            Intent intent = new Intent(getApplicationContext(), GuideVoluSettingActivity.class);
+            intent.putExtra("CLICKED_VOLU_KEY", clickedRowName);
+            intent.putExtra("CLICKED_VOLU_ID", guide.getMyVolunteers().get(clickedRowName));
+            startActivity(intent);
+            overridePendingTransition(0, 0);
+            return true;
+        }
+        else if (item.getItemId() == R.id.open_form_to_volunteer) {
+            Intent intent = new Intent(getApplicationContext(), GuideFormsPermissionActivity.class);
+            intent.putExtra("CLICKED_VOLU_KEY", clickedRowName);
+            intent.putExtra("CLICKED_VOLU_ID", guide.getMyVolunteers().get(clickedRowName));
+            startActivity(intent);
+            overridePendingTransition(0, 0);
+            return true;
+        }
+        return false;
     }
 
     private void setGuideName() {
